refactor(general): Placed server.js in api folder and changed start script
