+++ index.js
@@ -1,4 +1,4 @@
-import server from "./server.js";
+import server from "./api/server.js";
 
 const PORT = process.env.PORT || 3333;
